+++ tests/unit/api/v1/info/__init__.py
@@ -11,23 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-__all__ = ["NotFoundError", "NotUniqueError", "AuthenticationError"]
-
-
-class NotFoundError(Exception):
-    """Custom Argilla not found error. Use it for situations where an Argilla domain entity has not be found on the system."""
-
-    pass
-
-
-class NotUniqueError(Exception):
-    """Custom Argilla not unique error. Use it for situations where an Argilla domain entity already exists violating a constraint."""
-
-    pass
-
-
-class AuthenticationError(Exception):
-    """Custom Argilla unauthorized error. Use it for situations where an request is not authorized to perform an action."""
-
-    pass